refactor(net): Move wireless in its own file
Prepare for different implementations of wireless_network
