+++ src/CMakeLists.txt
@@ -23,6 +23,7 @@ endif()
 if(NOT ENABLE_NETWORK)
   list(REMOVE_ITEM files modules/network.cpp)
   list(REMOVE_ITEM files adapters/net.cpp)
+  list(REMOVE_ITEM files adapters/net_iw.cpp)
 endif()
 if(NOT ENABLE_I3)
   list(REMOVE_ITEM files modules/i3.cpp)
